% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_array_pdata.R
\name{intersect_array_pdata}
\alias{intersect_array_pdata}
\title{intersect_array_pdata}
\usage{
intersect_array_pdata(
  arraydata,
  pdata,
  genelist = NA,
  pdata_row2col = "sampleid"
)
}
\arguments{
\item{arraydata}{expression matrix}

\item{pdata}{pheondata data.frame}

\item{genelist}{a vector, with gene names}

\item{pdata_row2col}{transform the row.names of pdata into a col,default as sampleid}
}
\value{
list
}
\description{
intersect select cells and interested genes
}
\examples{
arraydata <- readRDS("~/sell/dev/cd8_all_array.RDS")
pdata <- readRDS("~/sell/dev/cd8_all_meta.RDS")
newdata <- intersect_array_pdata(arraydata,pdata)
}
