% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sell-package.R
\docType{package}
\name{sell-package}
\alias{sell}
\alias{sell-package}
\title{sell: Sell}
\description{
An R package in development for Single cELL analysis by using R6 object.
}
\author{
\strong{Maintainer}: Yanhua Zheng \email{yhzheng@cmu.edu.cn} (\href{https://orcid.org/0000-0003-1103-9579}{ORCID})

}
\keyword{internal}
