% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sell_melt.R
\name{sell_melt}
\alias{sell_melt}
\title{sell_melt}
\usage{
sell_melt(intersect_list, pdata_row2col = "sampleid")
}
\arguments{
\item{intersect_list}{a list object preprocessed by intersect_array_pdata function, which should has 2 items including arraydata and pdata}

\item{pdata_row2col}{transform the row.names of pdata into a col,default as sampleid}
}
\value{
data.frame
}
\description{
melt the arraydata and pdata into a long table for following plotting works
}
\examples{
melt_df <- sell_melt(newdata)
}
