% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_api_detect.R
\name{seurat_api_detect}
\alias{seurat_api_detect}
\title{seurat_api_detect}
\usage{
seurat_api_detect()
}
\value{
num
}
\description{
detect seurat api version, if NULL, try to install seurat packages
}
\examples{
seurat_api_detect()
}
