% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_decoder.R
\name{seurat_decoder}
\alias{seurat_decoder}
\title{seurat_decoder}
\usage{
seurat_decoder(
  seurat_object,
  HasNormalized = FALSE,
  HasZscored = FALSE,
  PCA = FALSE,
  UMAP = FALSE,
  TSNE = FALSE
)
}
\arguments{
\item{seurat_object}{the seurat object}

\item{HasNormalized}{the seurat object has normalized when TRUE}

\item{HasZscored}{the seurat object has Zscored when TRUE}

\item{PCA}{the seurat object has performed PCA when TRUE}

\item{UMAP}{the seurat object has performed UMAP when TRUE}

\item{TSNE}{the seurat object has performed TSNE when TRUE}
}
\value{
list
}
\description{
extract data from the seurat container
}
\examples{
sell_list <- seurat_decoder(seurat_object)
}
