% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat_encoder.R
\name{seurat_encoder}
\alias{seurat_encoder}
\title{seurat_encoder}
\usage{
seurat_encoder(
  arraydata,
  pdata,
  project = "SeuratProject",
  min.cells = 3,
  min.features = 200,
  reduction_dims = 1:20,
  Normalized = FALSE,
  Demention = FALSE
)
}
\arguments{
\item{arraydata}{expression matrix}

\item{pdata}{pheondata data.frame}

\item{project}{project names}

\item{min.cells}{min.cells in initing a seurat object}

\item{min.features}{min.features in initing a seurat object}

\item{reduction_dims}{dims used in data demention reduction}

\item{Normalized}{apply normalized methods}

\item{Demention}{apply demention reduction methods}
}
\value{
seurat_object
}
\description{
encode our arraydata and pdata into a seurat object and perform a quick computaion
}
\examples{
seurat_object <- seurat_encoder(newdata[["arraydata"]],newdata[["pdata"]],min.cells = 0,
                           min.features = 0,reduction_dims = 1:5)
}
