% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bagendclass.R
\name{BagEndClass}
\alias{BagEndClass}
\title{BagEnd}
\value{
R6Class
}
\description{
In a hole in the ground there lived a lot of cells,lol.
}
\examples{
sellobject <- BagEndClass$new(arraydata = arraydata,pdata = pdata)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-BagEnd-new}{\code{BagEndClass$new()}}
\item \href{#method-BagEnd-info}{\code{BagEndClass$info()}}
\item \href{#method-BagEnd-clone}{\code{BagEndClass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BagEnd-new"></a>}}
\if{latex}{\out{\hypertarget{method-BagEnd-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BagEndClass$new(arraydata, pdata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{arraydata}}{expression matrix}

\item{\code{pdata}}{pheondata data.frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BagEnd-info"></a>}}
\if{latex}{\out{\hypertarget{method-BagEnd-info}{}}}
\subsection{Method \code{info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BagEndClass$info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-BagEnd-clone"></a>}}
\if{latex}{\out{\hypertarget{method-BagEnd-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BagEndClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
